Allow load if HTTP authentication is required
I noticed how the user is redirected to an error message if the site is
not launched from localhost. Perhaps we should provide a way to allow
this site to be used on a remote server provided that it is in a
directory requiring HTTP authentication?

According to PHP's page on HTTP authentication
<http://php.net/manual/en/features.http-auth.php>,
$_SERVER['REMOTE_USER'] should be set if external authentication (from a
.htaccess file or what have you) was needed to reach the page and
$_SERVER['PHP_AUTH_USER'] should be set if PHP's HTTP authentication
methods were used.

Let me know what you think...